Better typing and structure
